Moves PAM source to its own files, changes default PAM Config, adds ccls cache to gitignore
